requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
einops==0.8.0
filelock==3.15.4
flash-attn==2.6.1
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pillow==10.4.0
pdf2image>=1.17.0,<2.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sympy==1.13.0
timm==1.0.7
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.42.4
triton==2.3.1
typing_extensions==4.12.2
urllib3==2.2.2
